[front] - refactor: move MessageItem component to messages subfolder
 - MessageItem is now part of a subfolder to better organize conversation components
 - Adjusted the import path for MessageItem in ConversationViewer to reflect new location
