+++ front/components/assistant/conversation/messages/MessageItem.tsx
